[RENAME] Product createdAt, updatedAt 타입 변경 및 enum 폴더 분리
