+++ src/main/java/com/programmers/smrtstore/domain/keep/application/KeepService.java
@@ -11,11 +11,10 @@ import com.programmers.smrtstore.domain.keep.presentation.dto.res.CreateKeepResp
 import com.programmers.smrtstore.domain.keep.presentation.dto.res.DeleteKeepResponse;
 import com.programmers.smrtstore.domain.keep.presentation.dto.res.KeepRankingResponse;
 import com.programmers.smrtstore.domain.keep.presentation.dto.res.KeepResponse;
-import com.programmers.smrtstore.domain.product.domain.entity.Category;
+import com.programmers.smrtstore.domain.product.domain.entity.enums.Category;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
