+++ src/main/java/com/programmers/smrtstore/domain/keep/infrastructure/KeepRepositoryCustom.java
@@ -2,8 +2,7 @@ package com.programmers.smrtstore.domain.keep.infrastructure;
 
 import com.programmers.smrtstore.domain.keep.presentation.dto.res.KeepRankingResponse;
 import com.programmers.smrtstore.domain.keep.presentation.dto.res.KeepResponse;
-import com.programmers.smrtstore.domain.product.domain.entity.Category;
-
+import com.programmers.smrtstore.domain.product.domain.entity.enums.Category;
 import java.util.List;
 
 public interface KeepRepositoryCustom {
