+++ src/main/java/com/programmers/smrtstore/domain/keep/presentation/dto/req/FindKeepByCategoryRequest.java
@@ -1,6 +1,6 @@
 package com.programmers.smrtstore.domain.keep.presentation.dto.req;
 
-import com.programmers.smrtstore.domain.product.domain.entity.Category;
+import com.programmers.smrtstore.domain.product.domain.entity.enums.Category;
 import lombok.Builder;
 import lombok.Getter;
