+++ src/main/java/com/programmers/smrtstore/domain/product/application/dto/req/UpdateProductOptionRequest.java
@@ -1,6 +1,6 @@
 package com.programmers.smrtstore.domain.product.application.dto.req;
 
-import com.programmers.smrtstore.domain.product.domain.entity.OptionTag;
+import com.programmers.smrtstore.domain.product.domain.entity.enums.OptionTag;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
