+++ src/main/java/com/programmers/smrtstore/domain/product/application/dto/req/UpdateProductRequest.java
@@ -1,6 +1,6 @@
 package com.programmers.smrtstore.domain.product.application.dto.req;
 
-import com.programmers.smrtstore.domain.product.domain.entity.Category;
+import com.programmers.smrtstore.domain.product.domain.entity.enums.Category;
 import java.net.URL;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
