+++ src/main/java/com/programmers/smrtstore/domain/product/application/dto/res/UpdateProductResponse.java
@@ -1,10 +1,10 @@
 package com.programmers.smrtstore.domain.product.application.dto.res;
 
-import com.programmers.smrtstore.domain.product.domain.entity.Category;
 import com.programmers.smrtstore.domain.product.domain.entity.Product;
+import com.programmers.smrtstore.domain.product.domain.entity.enums.Category;
 import java.net.URL;
-import java.sql.Timestamp;
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -25,8 +25,8 @@ public class UpdateProductResponse {
     private boolean availableYn;
     private boolean optionYn;
     private LocalDate releaseDate;
-    private Timestamp createdAt;
-    private Timestamp updatedAT;
+    private LocalDateTime createdAt;
+    private LocalDateTime updatedAT;
     private boolean discountYn;
